[Unfinished]NO_PYTHON has been removed from ir.cpp using function pointers, two pointer-initializers need to be created: *printPyObjectPtr; *pythonOpNamePtr;
