+++ setup.py
@@ -630,6 +630,7 @@ main_sources = [
     "torch/csrc/jit/init.cpp",
     "torch/csrc/jit/interpreter.cpp",
     "torch/csrc/jit/ir.cpp",
+    "torch/csrc/jit/ir_python.cpp",
     "torch/csrc/jit/fusion_compiler.cpp",
     "torch/csrc/jit/graph_executor.cpp",
     "torch/csrc/jit/python_ir.cpp",
